Project 2
Java a program that calculates the all items total, the discount amount, the subtotal, the sales tax amount, and the grand total for the department store. The program asks the sales associate to enter the cost of an item repeatedly until he/she enters zero to terminate the data entries. The associate should enter the current sales tax rate and the promotion code for the customer as well.